+++ src/main/java/rentalstore/statement/Statement.java
@@ -21,7 +21,7 @@ public class Statement extends BaseStatement {
     }
 
     @Override
-    public String getFooterString(double totalAmount, int frequentRenterPoints) {
+    public String getFooterString(double totalAmount, double frequentRenterPoints) {
         return "Amount owed is " + String.valueOf(totalAmount) + "\n"
                 + "You earned " + String.valueOf(frequentRenterPoints) + " frequent renter points";
     }
